// src/api/model/brandDetailsDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PizzeriaSimpleDto } from './pizzeriaSimpleDto';

export interface BrandDetailsDto {
  /** @nullable */
  logo?: string | null;
  pizzerias?: PizzeriaSimpleDto[];
  id?: string;
  name: string;
}